// src/app/services/clients/profile.client.ts
export interface ChangeNameRequest {
    publicID?: string | undefined;
    newName: string;
}