// src/app/services/clients/user.client.ts
export interface User {
    id: number;
    publicID?: string | undefined;
    email: string;
    password: string;
    registrationDate: Date;
    role?: string | undefined;
    deactivated: boolean;
}

export interface RegisterUserRequest {
    name?: string | undefined;
    dob: Date;
    password?: string | undefined;
    email?: string | undefined;
    role?: string | undefined;
}

export interface ChangePasswordRequest {
    publicID?: string | undefined;
    oldPassword: string;
    newPassword: string;
}